cleanup kwargs
