+++ starlite/kwargs/__init__.py
@@ -1,0 +1,3 @@
+from .kwargs_model import KwargsModel
+
+__all__ = ("KwargsModel",)
